Reset if we get confused. Buffer debugging
Count the number of clocks we've had without sending anything. Reset if this reaches 100. Buffer the debugging output to make sure we don't introduce latency to critical operations. Refactor the writeSerial1Debuggable method to allow sending blocks to reduce the function call overhead. Add new debugging flag for timing output as it's very annoying for most debugging. Use both bytes of the aperture in test messages.